chore: refactor useSwapAndBridgeContext
